+++ python/tests/worker_tests.py
@@ -272,6 +272,46 @@ class TestWorker(unittest.IsolatedAsyncioTestCase):
         await parent_queue.close()
         await queue.close()
 
+    async def test_process_job_respecting_the_concurrency_set(self):
+        num_jobs_processing = 0
+        pending_message_to_process = 8
+        wait = 0.01
+        job_count = 0
+        queue = Queue(queueName)
+
+        async def process(job: Job, token: str):
+            nonlocal num_jobs_processing
+            nonlocal wait 
+            nonlocal pending_message_to_process
+            num_jobs_processing += 1
+            self.assertLess(num_jobs_processing, 5)
+            wait += 0.1
+            await asyncio.sleep(wait)
+            self.assertEqual(num_jobs_processing, min(pending_message_to_process, 4))
+            pending_message_to_process -= 1
+            num_jobs_processing -= 1
+
+            return None
+
+        for _ in range(8):
+            await queue.add("test", data={})
+
+        worker = Worker(queueName, process, {"concurrency": 4 })
+
+        completed_events = Future()
+
+        def completing(job: Job, result):
+            nonlocal job_count
+            if job_count == 7:
+                completed_events.set_result(None)
+            job_count += 1
+
+        worker.on("completed", completing)
+
+        await completed_events
+
+        await queue.close()
+        await worker.close()
 
 if __name__ == '__main__':
     unittest.main()