+++ src/classes/queue.ts
@@ -352,11 +352,12 @@ export class Queue<
    * dependencies.
    *
    * @param jobId - The id of the job to remove
+   * @param opts - Options to remove a job
    * @returns 1 if it managed to remove the job or 0 if the job or
    * any of its dependencies were locked.
    */
-  remove(jobId: string): Promise<number> {
-    return this.scripts.remove(jobId);
+  remove(jobId: string, { removeChildren = true } = {}): Promise<number> {
+    return this.scripts.remove(jobId, removeChildren);
   }
 
   /**
@@ -429,8 +430,11 @@ export class Queue<
   /**
    * Retry all the failed jobs.
    *
-   * @param opts - contains number to limit how many jobs will be moved to wait status per iteration,
-   * state (failed, completed) failed by default or from which timestamp.
+   * @param opts: { count: number; state: FinishedStatus; timestamp: number}
+   *   - count  number to limit how many jobs will be moved to wait status per iteration,
+   *   - state  failed by default or completed.
+   *   - timestamp from which timestamp to start moving jobs to wait status, default Date.now().
+   *
    * @returns
    */
   async retryJobs(
@@ -443,6 +447,21 @@ export class Queue<
         opts.count,
         opts.timestamp,
       );
+    } while (cursor);
+  }
+
+  /**
+   * Promote all the delayed jobs.
+   *
+   * @param opts: { count: number }
+   *   - count  number to limit how many jobs will be moved to wait status per iteration
+   *
+   * @returns
+   */
+  async promoteJobs(opts: { count?: number } = {}): Promise<void> {
+    let cursor = 0;
+    do {
+      cursor = await this.scripts.promoteJobs(opts.count);
     } while (cursor);
   }